+++ backend/src/main/java/com/github/mdeluise/plantit/botanicalinfo/BotanicalInfoCreator.java
@@ -7,4 +7,5 @@ public enum BotanicalInfoCreator {
     INATURALIST,
     GBIF,
     PLANTNET,
+    PERENUAL,
 }